Flask tried
Django has a lot built-in and is highly opinionated
Flask is easy to get up and running but will require much manual work
But the project goal is clear, I will work on Flask and consider moving to Django if needed.

Something about Flask's simplicity just appeals to me.
Kinda like when I tried to build a database with json, moved to sqlite, and appreciated firebase. There's a progression to it all